Change file structure
Move the string utility function to libc/.
Move the I/O driver to cpu/ as they are CPU dependent.
