+++ cpu/ports.c
