--- kernel/kernel.h
@@ -1,6 +1,0 @@
-#ifndef KERNEL_H
-#define KERNEL_H
-
-void user_input(char *input);
-
-#endif
