+++ kernel/kernel_main.h
@@ -1,0 +1,6 @@
+#ifndef KERNEL_MAIN_H
+#define KERNEL_MAIN_H
+
+void user_input(char *input);
+
+#endif
