--- kernel/utils.c
@@ -1,88 +1,0 @@
-#include <stdint.h>
-#include <stddef.h>
-#include "utils.h"
-
-void kmemcpy(uint8_t *source, char *dest, int nbytes)
-{
-	int i;
-	for (i = 0; i < nbytes; i++) {
-		*(dest + i) = *(source + i);
-	}
-}
-
-void kmemset(uint8_t *dest, uint8_t val, uint32_t len)
-{
-	uint8_t *temp = (uint8_t *)dest;
-	for (; len != 0; len--)
-		*temp++ = val;
-}
-
-void int_to_ascii(int n, char str[])
-{
-	int i, sign;
-	if ((sign = n) < 0)
-		n = -n;
-	i = 0;
-	do {
-		str[i++] = n % 10 + '0';
-	} while ((n /= 10) > 0);
-
-	if (sign < 0)
-		str[i++] = '-';
-	str[i] = '\0';
-
-	/* reverse the string to get correct ascii representation
-	 * of the int.
-     */
-	reverse(str);
-}
-
-void reverse(char s[])
-{
-	int c, i, j;
-	for (i = 0, j = strlen(s) - 1; i < j; i++, j--) {
-		c = s[i];
-		s[i] = s[j];
-		s[j] = c;
-	}
-}
-
-int strlen(char s[])
-{
-	int i = 0;
-	while (s[i] != '\0')
-		++i;
-	return i;
-}
-
-void backspace(char *s)
-{
-	int len = strlen(s);
-	s[len - 1] = '\0';
-}
-
-void append_string(char *orignal, char *to_append)
-{
-	int len_orignal = strlen(orignal);
-	int len_to_append = strlen(len_to_append);
-	for (int i = 0; i < len_to_append + 1; i++) {
-		orignal[len_orignal + i] = to_append[i];
-	}
-}
-
-void append(char s[], char c)
-{
-	int len = strlen(s);
-	s[len] = c;
-	s[len + 1] = '\0';
-}
-
-int strcmp(char s1[], char s2[])
-{
-	int i;
-	for (i = 0; s1[i] == s2[i]; i++) {
-		if (s1[i] == '\0')
-			return 0;
-	}
-	return s1[i] - s2[i];
-}
