+++ libc/function.h
@@ -1,0 +1,7 @@
+#ifndef FUNCTION_H
+#define FUNCTION_H
+
+/* supress the unused variable warning */
+#define UNUSED(x) (void)(x)
+
+#endif
